+++ graylog2-web-interface/src/views/components/export/CSVExportModal.tsx
@@ -31,8 +31,8 @@ import Widget from 'views/logic/widgets/Widget';
 import { Icon, Spinner } from 'components/common';
 import { Modal, Button } from 'components/graylog';
 import BootstrapModalWrapper from 'components/bootstrap/BootstrapModalWrapper';
-import CSVExportSettings from 'views/components/searchbar/csvexport/CSVExportSettings';
-import CSVExportWidgetSelection from 'views/components/searchbar/csvexport/CSVExportWidgetSelection';
+import CSVExportSettings from 'views/components/export/CSVExportSettings';
+import CSVExportWidgetSelection from 'views/components/export/CSVExportWidgetSelection';
 import CustomPropTypes from 'views/components/CustomPropTypes';
 import { MESSAGE_FIELD, SOURCE_FIELD, TIMESTAMP_FIELD } from 'views/Constants';
 import { ExportSettings } from 'views/components/ExportSettingsContext';
